Better memory handling for error logs with a new configuration setting assigned that asks if stack trace should be written. Added ability to mass-delete error logs. Memory checks when logs are loaded. Added 'www-headers' input setting that tells Wave to return response-related data as headers instead of part of the output array. Fixed some documentation inconsistencies.

+++ tools/debugger.php
@@ -14,7 +14,7 @@
  * @license    GNU Lesser General Public License Version 3
  * @tutorial   /doc/pages/guide_tools.htm
  * @since      1.0.0
- * @version    3.1.8
+ * @version    3.2.1
  */
 
 // This initializes tools and authentication
@@ -31,12 +31,23 @@ if(empty($_GET)){
 	}
 	$errorLogs=scandir('..'.DIRECTORY_SEPARATOR.'filesystem'.DIRECTORY_SEPARATOR.'errors'.DIRECTORY_SEPARATOR);
 	// If it found error messages of any kind
-	if(count($errorLogs)>2){
-		if($errorLogs[2]!='.empty'){
-			header('Location: debugger.php?error='.$errorLogs[2]);
+	foreach($errorLogs as $error){
+		if($error!='.empty' && is_file('..'.DIRECTORY_SEPARATOR.'filesystem'.DIRECTORY_SEPARATOR.'errors'.DIRECTORY_SEPARATOR.$error)){
+			header('Location: debugger.php?error='.$error);
 			die();
 		}
 	}
+} elseif(isset($_GET['error'],$_GET['alldone'])){
+	$errorLogs=scandir('..'.DIRECTORY_SEPARATOR.'filesystem'.DIRECTORY_SEPARATOR.'errors'.DIRECTORY_SEPARATOR);
+	// If it found error messages of any kind
+	foreach($errorLogs as $error){
+		if(is_file('..'.DIRECTORY_SEPARATOR.'filesystem'.DIRECTORY_SEPARATOR.'errors'.DIRECTORY_SEPARATOR.$error)){
+			unlink('..'.DIRECTORY_SEPARATOR.'filesystem'.DIRECTORY_SEPARATOR.'errors'.DIRECTORY_SEPARATOR.$error);
+		}
+	}
+	// User will be redirected back to initial site
+	header('Location: debugger.php');
+	die();
 } elseif(isset($_GET['error'],$_GET['done'])){
 	// If error is considered to be 'fixed' then it will be removed, if it exists
 	if(file_exists('..'.DIRECTORY_SEPARATOR.'filesystem'.DIRECTORY_SEPARATOR.'errors'.DIRECTORY_SEPARATOR.$_GET['error'])){
@@ -88,29 +99,48 @@ if(empty($_GET)){
 		if(empty($_GET)){
 			echo '<p class="bold">There are no outstanding errors</p>';
 		} elseif(isset($_GET['error'])){
-			echo '<h3 onclick="if(confirm(\'Are you sure?\')){ document.location.href=document.location.href+\'&done\'; }" class="red bold" style="cursor:pointer;">Click to delete this error log</h3>';
-			// Getting error data
-			$errorData=file_get_contents('..'.DIRECTORY_SEPARATOR.'filesystem'.DIRECTORY_SEPARATOR.'errors'.DIRECTORY_SEPARATOR.$_GET['error']);
-			$errorData=explode("\n",$errorData);
-			foreach($errorData as $nr=>$d){
-				if(trim($d)!=''){
-					$d=json_decode($d,true);
-					// Last element in error log is used as preview of the full error
-					$time=array_shift($d);
-					$summary=array_pop($d);
-					echo '<div class="border block">';
-						echo '<b>'.$time.'</b>';
-						echo '<pre>';
-						print_r($summary);
-						echo '</pre>';
-						echo '<span class="bold" style="cursor:pointer;" onclick="if(document.getElementById(\'error'.$nr.'\').style.display==\'\'){document.getElementById(\'error'.$nr.'\').style.display=\'none\';} else {document.getElementById(\'error'.$nr.'\').style.display=\'\';}">MORE DETAILS</span>';
-						echo '<div id="error'.$nr.'" style="display:none;">';
+			echo '<h5 onclick="if(confirm(\'Are you sure? This deletes all error log entries!\')){ document.location.href=document.location.href+\'&alldone\'; }" class="red bold" style="cursor:pointer;float:right;">or delete all error logs</h5>';
+			echo '<h3 onclick="if(confirm(\'Are you sure?\')){ document.location.href=document.location.href+\'&done\'; }" class="red bold" style="cursor:pointer;width:400px;">Click to delete this error log</h3>';
+			
+			// Getting error file size
+			$filesize=filesize('..'.DIRECTORY_SEPARATOR.'filesystem'.DIRECTORY_SEPARATOR.'errors'.DIRECTORY_SEPARATOR.$_GET['error']);
+			
+			// Attempting to get memory limit
+			$memory=ini_get('memory_limit');
+			if($memory){
+				$memory=iniSettingToBytes($memory);
+				// Getting the amount of bytes currently allocated
+				$memory=$memory-memory_get_usage();
+				// available memory is 'about half' of the actual memory, just in case the array creation takes a lot of memory
+				$memory=intval($memory/2);
+			}
+			
+			// Error is only shown if memory limit could not be gotten or is less than the file size
+			if(!$memory || $filesize<$memory){
+				$errorData=file_get_contents('..'.DIRECTORY_SEPARATOR.'filesystem'.DIRECTORY_SEPARATOR.'errors'.DIRECTORY_SEPARATOR.$_GET['error']);
+				$errorData=explode("\n",$errorData);
+				foreach($errorData as $nr=>$d){
+					if(trim($d)!=''){
+						$d=json_decode($d,true);
+						// Last element in error log is used as preview of the full error
+						$time=array_shift($d);
+						$summary=array_pop($d);
+						echo '<div class="border block">';
+							echo '<b>'.$time.'</b>';
 							echo '<pre>';
-							print_r($d);
+							print_r($summary);
 							echo '</pre>';
+							echo '<span class="bold" style="cursor:pointer;" onclick="if(document.getElementById(\'error'.$nr.'\').style.display==\'\'){document.getElementById(\'error'.$nr.'\').style.display=\'none\';} else {document.getElementById(\'error'.$nr.'\').style.display=\'\';}">MORE DETAILS</span>';
+							echo '<div id="error'.$nr.'" style="display:none;">';
+								echo '<pre>';
+								print_r($d);
+								echo '</pre>';
+							echo '</div>';
 						echo '</div>';
-					echo '</div>';
+					}
 				}
+			} else {
+				echo '<p class="bold">This error log is too large for PHP to handle. Are you using error tracing in your configuration file? If so, then it is recommended to turn that setting off as that will generate smaller error log files.</p>';
 			}
 		}
